requirements.txt
annotated-types
anyio
bcrypt
beanie
certifi
click
dnspython
email-validator
exceptiongroup
fastapi
h11
httpcore
httpx
idna
iniconfig
lazy-model
mongomock
mongomock-motor
motor
packaging
passlib
pluggy
pydantic
pydantic-settings
pydantic_core
PyJWT
pymongo
pytest
pytest-mock
sentinels
sniffio
starlette
toml
tomli
typing_extensions
uvicorn

